legacy support: DD4hep->IRT geometry conversion for Juggler
Since the geometry bindings between DD4hep and IRT are now maintained
here in EICrecon, this branch provides Juggler support, while not
introducing upstream breaking changes from the PID data model.
In particular, the bindings are at
```
src/services/geometry/rich/richgeo
```
EICrecon itself is not needed to be built; in practice this `richgeo`
directory is externally built to a standalone library.