Support for for Acts-20.0.0 in EICrecon
Most conversion are just the usual interface changes. The only one
that is a bit trickier is where we have to explicitly route this
through the explicity EICrecon shared_ptr interfaces. In those cases
we cannot simply rely on the internal consistency of Acts.

This now requires Acts 20.0.0 in find_package