feat: CMAKE_PROJECT_VERSION as git describe (e.g. v0.6.2-5-g5f598b8-dirty)
This sets the version number based on the output of `git describe` for
installations where git is installed and such output is available. This
notably excludes tar balls of releases. We are not currently using
those. A release procedure is required that addresses this, e.g. by
creating a `VERSION` file before release as part of a release workflow.