CalorimeterIslandCluster: autotune transverse shower profile parameter
The current implementation uses calorimeter cell size for value of shower
width parameter. This approach probably doesn't work for calorimeters other than the Imaging.
The width should be related to the Moli`ere radius for the material, but
possibly also the geometry (even location of the cluster in the calorimeter).
This commit implements a parameter scan that should help users to get get
cluster splitting working, while we are trying to figure out a better way to
handle this.