feat: documentation rework  (#533)
### Briefly, what does this PR introduce?
The EICrecon repo should not serve as the source of documentation for
unrelated components. It is not updated when those other components are
updated, and therefore likely wrong in several cases.

List of changes:
- Remove `edpm` since not a supported or sustainable project.
- Removed unsupported direct singularity calls to cvmfs since they tend
to fail without additional flags.
- Removed non-standard singularity installation instructions that pull a
`.deb` file from the internals of an apt dist server.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [x] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [x] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.