EcalEndcapP{,Insert}RawHits: revert back to unitless definitions (#584)
The MeV multipliers were introduced in #408, but they don't make sense
from the dimensional analysis point (terms are divided by sqrt(E), 1 and
E respectively). Each element of the array would need to have its own
unit, but the commonly accepted convention is to use fractions and
assume GeV. This commit restores that convention. The actual behavior in
reconstruction is not changed by this PR, because the values stay the
same.