EcalBarrelScFiRawHits: revert back to unitless definitions (#583)
The current value was introduced in
13b9c92188d9cacedc5c4b9ca0c80404c21f1c3e. There should have been three
values, and using units for those doesn't make sense from the
dimensional analysis point (terms are divided by sqrt(E), 1 and E
respectively). Each element of the array would need to have its own
unit, but the commonly accepted convention is to use fractions and
assume GeV. The actual behavior in reconstruction is not changed by this
PR, because the values stay the same.