BUGFIX: CalorimeterIslandCluster: compare absolute values for distance-based adjacency settings (#610)
Currently, if the adjacency is defined using distance-based options like
u_localDistXY, there will be cases when distance vector reported by
hitsDist will have negative components. When those components are
compared to positive numbers, this leads to an always-true neigbourhood
condition. The clusters may be wrongly merged because of that.