services: use virtual methods, like in Juggler (#640)
### Briefly, what does this PR introduce?

This includes a change needed for #605, separated for purposes of
focused review.
Unit testing could benefit from a proper isolation of the plugins and
services. Services implemented in Juggler define interfaces with virtual
method dispatching. This PR does a small step towards that scheme.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No

### Does this PR change default behavior?
No