Updated seed phi calculation and seed finder configuration (#641)
### Briefly, what does this PR introduce?
This PR updates the calculation of the azimuthal angles for the ACTS
seeds as discussed on slides 7-11 here:


https://indico.bnl.gov/event/18894/contributions/75735/attachments/47132/79949/ACTS%20seed%20reconstruction.pdf

Also, this PR updates the seed finder configuration parameters to the
most up-to-date values based Emma's and Rey's studies:


https://indico.bnl.gov/event/18894/contributions/75735/attachments/47132/79952/param%20exploration%20%282%29.pdf

Some default values for the seed filter are also set, but do not seem to
be used in the track seeding algorithm right now.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #621)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [x] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No
### Does this PR change default behavior?
No

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Wouter Deconinck <wdconinc@gmail.com>