CalorimeterHitDigi: remove u_ref (#660)
In the current code, `u_ref` is used to set value for bits of the cell
ID that were removed by `id_mask`. The resulting value `hid` is printed
out to logs and used as unique key, but never propagates anywhere.
Removing support to u_ref will have no effect on the reconstruction, and
will be equivalent to setting `u_ref` to `{0}`.

See discussion in
https://eic.cloud.mattermost.com/main/pl/d1c8cm1rabnrjnxqumbyioua7c on
how u_ref is part of a legacy logic that had to be disabled.