fix(dRICH): extrapolate quantum efficiency endpoints (#714)
### Briefly, what does this PR introduce?
Extrapolate endpoints of the quantum efficiency curve down to zero. Two
points are added:
| wavelength | quantum efficiency |
| --- | --- |
| 315 nm | 0.0 |
| 1000 nm | 0.0 |

This is a minor change, applied to both the default config
`PhotoMultiplierHitDigiConfig` and the dRICH-specific config in
`DRICH.cc` (along with a syntax simplification).

This goes along with other material property extrapolations from
https://github.com/eic/epic/pull/445

### What kind of change does this PR introduce?
- [x] Bug fix: this is a part of https://github.com/eic/epic/issues/439
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [x] Changes have been communicated to collaborators
  - @chchatte92 

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
no
### Does this PR change default behavior?
no, but `<NPE>` will increase slightly