Add reconstructed electron factory, algorithm utilizing E/p cut  (#751)
### Briefly, what does this PR introduce?
Provides electron candidates as a first step towards a DIS electron
finder.
This PR uses the recently provided MC particle to Reco cluster
associations to produce a `ReconstructedElectrons` collection. Electrons
candidates are identified with an E/p cut using Reco particle momentum
and Redo cluster energy (only from ECAL detectors for now).

Couple of comments/potential improvements for reviewers:
- I am not sure how best to use the Config objects to allow parameters
to be modified (e.g. min/max E/p in this case)
- AFAIK this creates a copy of the reconstructed particles - since this
(ReconstructedElectrons) is a strict subset of the
ReconstructedParticles collection, is there a way to store as such
(without duplicating)
- I am currently pulling from (by name) several locations (see
`reco.cc`) to get all possible source of
`edm4eic::MCRecoClusterParticleAssociation` - is there a way to instead
ask for all sources of a given type?


### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added: 
  - Not sure if new tests needed, 
- "ReconstructedElectrons" added to polio default output (include list)
- simply running on a recent sim file should suffice as a test -
expected behavior is to produce the "ReconstructedElectrpons"
collection. A second level of test is to ensure that the collection is
not empty (assuming at least some electrons within acceptance are in the
simulated input)
- [ ] Documentation has been added / updated
- [x] Changes have been communicated to collaborators: discussion in S&C
meeting on June 5

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No

### Does this PR change default behavior?
No

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>
Co-authored-by: Wouter Deconinck <wdconinc@gmail.com>