feat: more modern CMake by using targets instead of include_directories (#720)
### Briefly, what does this PR introduce?
This aims to modernize, simplify, and in some cases make more robust the
CMake configuration. Primarily this involves removing constructs like
`set(Acts_INCLUDE_DIR ${Acts_DIR}/../../../include)` which are prone to
change, and are not necessary when using targets such as `ActsCore`,
etc. This PR also removes a number of unused blocks of code.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue: cmake modernization)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.