feat: rewrite MC2SmearedParticle to operate on collections
This reworks the MC2SmearedParticle algorithm and factory to operate on
podio collections, which are passed to the algorithm. It moves the loop
over MCParticles into the algorithm (maybe at some point the factories
can parallelize this, but we're not there, and consistency in how
algorithms interact with collections is more important right now). It
also removes the (defunct) smearing configuration.