feat: convert calorimetry JFactoryT to JChainFactoryT
All of the calorimetry factories are currently JFactoryT instead of
JChainFactoryT, which means that they don't allow for providing a
configuration object in the plugin (think: BEMC.cc).

This converts all calorimetry JFactoryT factories to JChainFactoryT,
which requires that we specify also the set of input tags, and the
output tag in the JChainFactoryGeneratorT constructor in the plugin.
These are therefore removed from the factories since they are set in the
plugin.

The third argument of the JChainFactoryGeneratorT can become the config
object, which can now be introduced in a separate PR, algorithm by
algorithm.

All this fits in a three-step process towards simplification:
- make calorimetry algorithms accept external config objects
- collect all configuration in plugin
- remove or reduce dedicate factory code