fix: place truth vertex on perigee surface in truth seeding (#783)
### Briefly, what does this PR introduce?
This fixes the truth vertex issue introduced in #766 and discussed in
#779 by placing the truth vertex on the perigee surface with its
two-vector. This 'mostly' recovers the original behavior, but it
necessarily has differences because the truth vertex propagation is
modeled differently (but now identical to how this is done in realistic
seeding).

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [x] Changes have been communicated to collaborators @veprbl
@bschmookler

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
Yes, some changes in truth seeded tracks to be expected.