Added CKF tracking with real seeds. (#711)
### Briefly, what does this PR introduce?
This introduces factories which allow the output of the realistic seeder
to be fed into the CKF tracking algorithm. The implementation is
explained on slide 6
[here](https://indico.bnl.gov/event/19363/contributions/77040/attachments/47759/80987/track_recon_060123.pdf).
A factory output called _CentralCKFSeededTrajectories_ is created.
However, these real seeded ACTS trajectories are not currently being
sent as input to the factories downstream which write out the EDM4EIC
output which then gets written to the standard EICRecon output TTree. If
possible, we can add a user-controllable flag to switch between using
the truth-seeded and real-seeded trajectories in those factories.

The _CentralCKFSeededTrajectories_ data can be currently accessed using
an EICRecon Plugin. When using such a Plugin, the results of simulating
10,000 single muon events are shown on slides 7-9
[here](https://indico.bnl.gov/event/19363/contributions/77040/attachments/47759/80987/track_recon_060123.pdf).

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No

### Does this PR change default behavior?
No

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Wouter Deconinck <wdconinc@gmail.com>