feat: rewrite MC2SmearedParticle to operate on collections (#778)
### Briefly, what does this PR introduce?
This reworks the MC2SmearedParticle algorithm and factory to operate on
podio collections, which are passed to the algorithm. It moves the loop
over MCParticles into the algorithm (maybe at some point the factories
can parallelize this, but we're not there, and consistency in how
algorithms interact with collections is more important right now). It
also removes the (defunct) smearing configuration.

This also removes the `generatorStatus == 0` entries as done in #772.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue: use collections in MC2SmearedParticles)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.