fix: rm failing MC2SmearedParticleConfig include
This removes an include of a removed header.

When we run CI tests, we use an eic-shell container that has EICrecon already installed. That means that removing a header file in the test build can still cause that header file to be found in the system directories. This then only shows up in an actual clean build later on.