feat: settable seed per CalorimeterHitDigi algorithms
This adds a settable seed (uint32_t) for the CalorimeterHitDigi
algorithm, defaulting to zero, but settable in the derived factories. If
we move to a different structure of factories, they should be able to
produce the exact same output, per algorithm/factory, if given the same
seed. Not sure this really gives us that much more insight or control,
but it gives us a way to make sure we can control each algorithm
independently.