fix: rm unused and unmaintained pyjano (#813)
### Briefly, what does this PR introduce?
It is possible that at some point we aim to steer eicrecon or jana from
python. In that case we can revive this. At this point, the pyjano
'test' isn't a test, isn't maintained, and doesn't have sufficient
information for users or developers to use.

### What kind of change does this PR introduce?
- [x] Bug fix (issue: debt)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.