fix: rm unimplemented csh support (#816)
### Briefly, what does this PR introduce?
We install a placeholder `eicrecon-this.csh` which doesn't do anything,
and this leads to a degraded user experience on csh-based systems, such
as the JLab ifarm where users might think `source bin/eicrecon-this.csh`
can be used. Yes, I know it prints a warning.

We could also implement the csh script, but no one has felt strongly
enough about that in the past year to even mention it, as far as I know.
It would add a need to maintain duplicate environment scripts. Until
there is some significant demand for this, we should remove the
unimplemented script, I think.

### What kind of change does this PR introduce?
- [x] Bug fix (issue: csh support not implemented)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.