feat: rm utility StringHelpers.h from which only one function ever used
`StringHelpers.h` was copied into EICrecon with many functions, but only
one function is ever used (`ReplaceAll`), and it is used many times for
the same purpose (chopping the ".so" off the plugin name)

However, JANA2 already removes the ".so" as of 2.0.8
(https://github.com/JeffersonLab/JANA2/commit/8ed0703071c8f00ea7e41cfd2a1e080d58bf9743),
and this puts platform specific code where it doesn't belong (".so" will
not work for Mac, for example).

Keeping `StringHelpers.h` around adds to the maintenance burden,
especially since the other functions are not explicitly or implicitly
tested by use in code.