fix: convert ImagingPixelReco to use CalorimeterHitReco (#821)
### Briefly, what does this PR introduce?
This replaces the ImagingPixelReco with the more broadly applicable
CalorimeterHitReco (and removes ImagingPixelReco), after discussion with
@Chao1009. That fixes the time resolution mismatches discussed in #807.

At some point CalorimeterHitDigi and CalorimeterHitReco will learn ToT
and other features of modern digitization (including AstroPix) and then
we can update the parameters here. For now, this uses a 'simple' TDC
assumption with bit resolution of 3.25 ns, after discussion with
@mariakzurek.

After this PR is applied to `imaging-cluster-reco-collections` in #807,
the differences for a single event (first array is main vs second array
is this branch) are
```
EcalBarrelImagingClusters.time
NOT EQUAL
[[0.00033, 0.00034, 0.000649]]
[[3.25, 3.25, 6.5]]
EcalBarrelImagingMergedClusters.time
NOT EQUAL
[[3.87, 6.79, 3.64]]
[[3.25, 6.5, 3.25]]
```
i.e the timing for EcalBarrelImagingClusters is now sensible instead of
incorrectly of picosecond scale, and it is indeed digitized with 3.25
bit resolution (as it happens, tracks take about 3.5 ns to reach the
detector, or at least 2.5 ns for photons at eta = 0). That time is now
used for the merged clusters.

### What kind of change does this PR introduce?
- [x] Bug fix (issue: incorrect hit time reconstruction in imaging
pixels)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [x] Changes have been communicated to collaborators @Chao1009
@mariakzurek @sly2j

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.