fix: rename imaging/scfi module to sector
As of https://github.com/eic/epic/pull/490, the naming of the sectors of the imaging and scfi is now correct in the readout field. This is reqiured because there are now actual modules implemented (i.e. chip modules placed on staves). The hierarchy/nomenclature defined by the DSC is system > sector > layer > stave > module > pixel (x,y).