fix: allow for = in working dir for datamodel glue
We specify the working dir where datamodel glue headers are
written as a key-value pair argument, `WORKING_DIR=dir`. If
we then take field 1 after a split on '=', we discard parts
of the directory if it contains '=' characters. This fixes
the split to only split on the first '=' character.

Use case: when we build a git tag or branch in spack, we
give an equivalent version for dependency resolution
purposes, e.g `eicrecon@git.feature-branch=v1.4.0`. For
the version of spack we are using, this leads to staging
directories with that full version string.