fix: allow for = in working dir for datamodel glue (#844)
### Briefly, what does this PR introduce?
We specify the working dir where datamodel glue headers are written as a
key-value pair argument, `WORKING_DIR=dir`. If we then take field 1
after a split on '=', we discard parts of the directory if it contains
'=' characters. This fixes the split to only split on the first '='
character.

Use case: when we build a git tag or branch in spack, we give an
equivalent version for dependency resolution purposes, e.g
`eicrecon@git.feature-branch=v1.4.0`. For the version of spack we are
using, this leads to staging directories with that full version string.

### What kind of change does this PR introduce?
- [x] Bug fix (issue: [build
failures](https://eicweb.phy.anl.gov/containers/eic_container/-/jobs/1840504))
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.