fix: rm unused output{Seeded}Particles from TrackingResult_factory
This removes the unused output{Seeded}Particles collection from the
tracking plugin, and removes the ReconstructedParticle output from the
TrackingResult algorithm and factory.

The construction of ReconstructedParticles (which includes PID
information) is not something that should be done at this stage of the
track reconstruction chain (and indeed the results are ignored).