services/log: remove timestamp from log messages, allow configuring format (#851)
During development, it is often needed to diff log files, which can be
cumbersome if timestamps are printed. It doesn't seem like we have a
strong need to get timestamps in the production logs at the time, but if
that will be needed, it can now be re-enabled using eicrecon::LogFormat
option.