Realistic scfi/imaging energy/position cluster merger (#822)
### Briefly, what does this PR introduce?
This adds the realistic scfi/imaging energy/position cluster merger to
the default reconstruction. It also renames the truth energy/position
cluster merger collections and associations to end with
`..TruthClusters`.

There is a bit of a design decision here that was made: if the matched
energy and position cluster disagree on the MCParticle entry, then two
associations are written, one for each possible entry, and at half
weight.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [x] Changes have been communicated to collaborators: @Chao1009
@mariakzurek @sly2j

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.