feat: produce edm4eic::TrajectoryCollection in CKFTracking
This moves the creation of podio output types into the CKFTracking
algorithms (it should accept and produce only podio collections), with
the assumptions from the previous ParticlesFromTrackFit algorithm. No
changes expected (with the exception of changed collection IDs due to
the addition of a new output collection).

The CKFTracking algorithm now produces both the desired podio output
type and the deprecated ActsExamples::Trajectories, which is still used
by some downstream algorithms. Next is to migrate those algorithms over.

The TrackingResult factory and ParticlesFromTrackFit algorithms are now
obsolete and removed.