fix: restore merging of hits in TrackerHitCollector
This restores merging of all tracker hits in the TrackerHitCollector. It
does not assume that all collections are present but empty if not
supported by the geometry configuration. Instead, it catches the
exception on JEvent::GetCollectionBase if a podio collection is
requested that does not exist, and continues with the other (existing)
podio collections. The transition to podio collections upstream, in

This also reworks the factory to the new multifactory approach, moves it
into src/factories, and uses the modular algorithms for the operations
on podio collections instead of replicating the functionality in the
factory.