fix: restore merging of hits in TrackerHitCollector (#868)
### Briefly, what does this PR introduce?
This restores merging of all tracker hits in the TrackerHitCollector. It
does not assume that all collections are present but empty if not
supported by the geometry configuration. Instead, it catches the
exception on JEvent::GetCollectionBase if a podio collection is
requested that does not exist, and continues with the other (existing)
podio collections. The transition to podio collections upstream, in

This also reworks the factory to the new multifactory approach, moves it
into src/factories, and uses the modular algorithms for the operations
on podio collections instead of replicating the functionality in the
factory.

### What kind of change does this PR introduce?
- [x] Bug fix (issue: #840 introduced a regression by removing all hits
from CentralTrackingRecHits)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
Yes, ensures CentralTrackingRecHits are filled again.