CalorimeterHitDigi: don't add a hit when none of the merged hits meet the timing requirement (#856)
There was an edge case when all hits could be skipped in the previous
loop and then the algorithm would still proceed to add a hit with
edep=0. Most common case to trigger that, would be if there is only a
single hit (e.g. if we were to use this code path for calorimeters for
which the hit merging is not required).