feat: use podio collections in tracking digi/reco; rm unused interfaces (3rd attempt) (#872)
### Briefly, what does this PR introduce?
This converts SiliconTrackerDigi (used for MPGDs too) and
TrackerHitReconstruction to processing podio collections instead of
individual hits. It adapts their factories to multifactories, and
updates the plugins. It removes now unused algorithm interface
definitions (IObjectProducer, etc) and the remaining duplicated and
unused TruthTrackSeeding algorithm.

Follows #840 and #871.

### What kind of change does this PR introduce?
- [x] Bug fix (issue: remove unused algorithms, interfaces, and
factories)
- [x] New feature (issue: move tracking digitization and reconstruction
to podio collections)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No. This should lead to identical output as before..

### Does this PR change default behavior?
No.