feat: produce edm4eic::TrajectoryCollection in CKFTracking  (#860)
### Briefly, what does this PR introduce?
This moves the creation of podio output types into the CKFTracking
algorithms (it should accept and produce only podio collections), with
the assumptions from the previous ParticlesFromTrackFit algorithm. No
changes expected (with the exception of changed collection IDs due to
the addition of a new output collection).

The CKFTracking algorithm now produces both the desired podio output
type and the deprecated ActsExamples::Trajectories, which is still used
by some downstream algorithms. Next is to migrate those algorithms over.

The TrackingResult factory and ParticlesFromTrackFit algorithms are now
obsolete and removed.

TODO:
- [x] merge #859 first, it is included here

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.