CalorimeterHitDigi: unified code path for digitization with and without the hit merging (#864)
### Briefly, what does this PR introduce?
This should be easier to maintain. We probably could even refactor
merging into
a separate algorithm, if we wanted to.

A side effect of this is that non-merging digitization now reorders hits
(the
cell id is used as a key for an std::unordered_map).

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [x] Changes have been communicated to collaborators @Chao1009 

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No

### Does this PR change default behavior?
Yes, the hits will be reordered.