Changed kinematic calculations to use edm4eic vector utils in ParticlesWithPID (#914)
### Briefly, what does this PR introduce?
The std::hypot method used to define p_mag in ParticlesWithPID.cc by
default returns a float. This is insufficient precision when looking at
particles travelling very close to the beam axis.

This PR changes to using the edm4eic vector utils methods which return
doubles.

Additionally the eta and phi tolerance values provided to match the MC
and reconstructed particles do not make sense for the particles along
the beamline so if both are below -5 they are assumed to match.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
no
### Does this PR change default behavior?
no

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>