fix: move Eigen out of SpdlogFormatters (#920)
### Briefly, what does this PR introduce?
One of the largest source of slow builds is the inclusion of `Eigen`
headers indirectly in many source files. This removes one of these
instances. Rather than including and using `Eigen::MatrixBase<T>` in the
SpdlogFormatters header, this moves the 'expensive' formatters to the
source files where they are used.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #628 but doesn't fix it all)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>