fix: remove non-POD-ness from OrthogonalTrackSeedingConfig
The OrthogonalTrackSeedingConfig is now fully POD, which means that
including the configuration does not incur the wrath/weight of all
of Acts. The TrackSeeding algorithm now takes care of turning POD
configs into the Acts config objects (much like how this is done in
the other tracking algorithms, though the orthogonal seeding takes
it to a new level of course).

Only necessary headers are now included in the POD config, and the
headers in TrackSeeding.h and TrackSeeding.cc were reevaluated as
well due to this shuffle. Two future FIXMEs have been added since
the older Acts version we use has missing includes that have since
been added. Check back when we get to Acts 23.0.0 and 27.2.0.

No output collection changes at all expected due to this PR.