fix(ci): don't use pre-filled JANA_PLUGIN_PATH with EICrecon plugins (#942)
### Briefly, what does this PR introduce?
`JANA_PLUGIN_PATH` in eic-shell (and therefore CI environments) is
prepopulated with `/usr/local/lib/EICrecon/plugins:/usr/local/plugins`.
This means that removing (intentionally or otherwise) a plugin in a PR
will still cause it to be found in `/usr/local/lib/EICrecon/plugins`. We
need `/usr/local/plugins` though, because that contains the `janadot`
plugin to create callgraphs.

There's an issue reported to JANA2
(https://github.com/JeffersonLab/JANA2/issues/136) that points out that
installing plugins in a shared namespace at `/usr/local/plugins` may
lead to issues and does not conform to the linux FHS. But at least
EICrecon plugins are not installed there, otherwise we would have
exactly this issue.

### What kind of change does this PR introduce?
- [x] Bug fix (issues, potentially)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.