fix: enable rpath on build and install linking (#938)
### Briefly, what does this PR introduce?
Enable rpath/runpath linking. The goal of this PR is to add the
dependency libraries (JANA, podio, EDM4hep, EDM4eic) with their paths
into rpath and runpath, and to have all internal library dependencies be
explicit as well. Ultimately, no LD_LIBRARY_PATH should need to be set.

And the same on MacOS...

### What kind of change does this PR introduce?
- [x] Bug fix (issue #842)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.

---------

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>