fix(ci): use npsim instead of ddsim (#955)
### Briefly, what does this PR introduce?
We are trying to exercise the whole pipeline, but not throwing any
optical photons. This means the dRICH does not currently see anything
and we don't test the reconstruction and particle matching.

This is expected to result in 1) changes, 2) slower simulation running,
3) dragons.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.