fix: use RPATH in build tree, unset LD_LIBRARY_PATH for ctest (#978)
### Briefly, what does this PR introduce?
This will (hopefully) fix the ctest failures in CI. We were not explicit
about how we want rpath to be handled in the build tree (now we are),
but more importantly on binaries the rpath settings in CMake set the
RUNPATH elf record, which is lower precedence than the entries in
LD_LIBRARY_PATH. That means that any executable (such as the test
executable) will still defer to LD_LIBRARY_PATH, if set. So, we unset it
for ctest.

This doesn't affect our other executable, eicrecon, because that isn't
actually linked against anything (or nothing that is internal to our
source tree). Instead it uses the dlopen to plugins explicitly in
JANA_PLUGIN_PATH. When those are opened, the rpath to their dependencies
is respected.

### What kind of change does this PR introduce?
- [x] Bug fix (issue, see also #976)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.