fix: avoid undefined unsigned integer behavior in pid size checks
This avoids undefined behavior flagged by ubsan in the statement
`std::size_t n_tracks = -1;` and in the signed/unsigned integer
comparisons in the pid system, which caused at least MergeTracks to
always fail. It also adds output on the problematic sizes, e.g.
```
[DRICH:DRICHMergedTracks] [error] cannot merge input track collections
with different sizes 3, 5
[DRICH:DRICHIrtCherenkovParticleID] [error] radiators have differing
numbers of TrackSegments 3, 5, 0
```