feat: std::optional in cluster reco CoG (#1000)
### Briefly, what does this PR introduce?
In CalorimeterClusterRecoCoG, we use a helper function which can
sometimes decide not to return a cluster. `std:: optional` makes that
intent clearer.

See discussion in
https://github.com/eic/EICrecon/pull/989#discussion_r1330628992.
Suggestion is to merge #989, merge this, then put the actual `return {}`
for zero weights in a third PR.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.