JFactoryPodioT.h: don't populate mData with clones of collection objects
We are observing inconsistent crashes in delete operator invocation in
ClearData(). These faults occur on macOS with a high rate, specifically for
CalorimeterHitContribution. We also have reports of some percent of
reconstruction jobs (on Linux) crashing specifically on ReconstructedParticle.

There might be an issue with PODIO object clones, they should be pointing to
the same object data. There doesn't seem to be any facility for reference
counting.

It is not clear if mData is used or just populated for compatibility. This is
an attempt at disabling the functionality in order to mitigate the issue.